Refactor into namespaces
Refactored Step-related entities to Steps namespace and
StepDefinition-related entities to StepDefinitions namespace.
