+++ src/Libraries/Gauge.NCrunch.Runner/Gauge.NCrunch.Runner.csproj
@@ -42,18 +42,19 @@
     </Compile>
     <Compile Include="CodeContracts\Contract.cs" />
     <Compile Include="IScenarioRunner.cs" />
-    <Compile Include="IStep.cs" />
-    <Compile Include="IStepAttributedMethodResolver.cs" />
-    <Compile Include="IStepDefinition.cs" />
-    <Compile Include="IStepDefinitionFactory.cs" />
-    <Compile Include="IStepDefinitionResolver.cs" />
-    <Compile Include="IStepFactory.cs" />
-    <Compile Include="IStepMatcher.cs" />
-    <Compile Include="MultipleMatchingStepDefinitionsException.cs" />
+    <Compile Include="Steps\IStep.cs" />
+    <Compile Include="StepDefinitions\IStepAttributedMethodResolver.cs" />
+    <Compile Include="StepDefinitions\IStepDefinition.cs" />
+    <Compile Include="StepDefinitions\IStepDefinitionFactory.cs" />
+    <Compile Include="StepDefinitions\IStepDefinitionResolver.cs" />
+    <Compile Include="Steps\IStepFactory.cs" />
+    <Compile Include="StepDefinitions\IStepMatcher.cs" />
+    <Compile Include="StepDefinitions\MultipleMatchingStepDefinitionsException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScenarioRunner.cs" />
-    <Compile Include="StepDefinitionNotFoundException.cs" />
-    <Compile Include="StepDefinitionResolver.cs" />
+    <Compile Include="StepDefinitions\StepDefinitionNotFoundException.cs" />
+    <Compile Include="StepDefinitions\StepDefinitionResolver.cs" />
+    <Compile Include="Steps\StepFactory.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
