+++ src/Libraries/Gauge.NCrunch.Runner/StepDefinitions/IStepAttributedMethodResolver.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Reflection;
 
-namespace Gauge.NCrunch.Runner
+namespace Gauge.NCrunch.Runner.StepDefinitions
 {
     public interface IStepAttributedMethodResolver
     {
