+++ src/Libraries/Gauge.NCrunch.Runner/StepDefinitions/IStepDefinitionFactory.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
 
-namespace Gauge.NCrunch.Runner
+namespace Gauge.NCrunch.Runner.StepDefinitions
 {
     public interface IStepDefinitionFactory
     {
