+++ src/Tests/Gauge.NCrunch.Runner.Tests/Gauge.NCrunch.Runner.Tests.csproj
@@ -89,9 +89,9 @@
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScenarioRunnerTests.cs" />
-    <Compile Include="MultipleMatchingStepDefinitionsExceptionTests.cs" />
-    <Compile Include="StepDefinitionNotFoundExceptionTests.cs" />
-    <Compile Include="StepDefinitionResolverTests.cs" />
+    <Compile Include="StepDefinitions\MultipleMatchingStepDefinitionsExceptionTests.cs" />
+    <Compile Include="StepDefinitions\StepDefinitionNotFoundExceptionTests.cs" />
+    <Compile Include="StepDefinitions\StepDefinitionResolverTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Libraries\Gauge.NCrunch.Runner\Gauge.NCrunch.Runner.csproj">
@@ -102,6 +102,9 @@
   <ItemGroup>
     <None Include="app.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Steps\" />
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
